refactor: extract event loop to worker_eventloop.go
